requirements.txt
appdirs==1.4.3
Django==1.10.6
djangorestframework==3.6.2
ordereddict==1.1
packaging==16.8
pyparsing==2.2.0
requests==2.13.0
six==1.10.0
WooCommerce==1.2.1
